Added Edit Task funcitonality. Added mobile interaction.
 Updated database.update_task and app.py edit_task to properly edit task details and perform appropriate edits to the subsequent task instance. When name, description, points or time are changed, only the task is updated. When the frequency, start or end date are changed all pending task instances of that task are deleted are new task instances are created that follow the new schedule. Only task instances after and including the current day are deteled, and if the new start date is before the current date, the current date is used instead.
Added task_scheduler.py and moblie_interaction.py which allow the application to send mobile notifications when a task is due and 5 minutes beforehand using Pushbullet. task_scheduler runs a seperate thread from __main__ which checks all of today's tasks every minute to see if they are due or are due in 5 minutes. When a task is due, it uses the send_task_notifiation in mobile_interaction to send a notification about the task. Pushbullet requires an authentication key that can be set in .env